refactor: extract L1 table types into meta/l1.rs
Move L1Entry, L1Table, their trait impls and the test_l1_table unit
test into a new meta/l1.rs submodule. The test moves with the code
because it constructs L1Entry via its private tuple field, which is
only reachable from within the type's own module. No logic changes.
